Moved Login page into own component (Login.jsx), added login credentials to docker-compose template, added auth to GET /api/ui/config
