+++ docker-compose.yaml
@@ -17,6 +17,9 @@ services:
       # - /path/to/cookies.txt:/opt/explo/cookies.txt # Path to optional cookies file (for yt-dlp)
     environment:
       - TZ=UTC # Change this to the timezone set in ListenBrainz (default is UTC)
+      # Web UI credentials (required for login)
+      - UI_USERNAME=
+      - UI_PASSWORD=
 
       # [Legacy] Schedules are managed through the web UI.
       # These can still be set here for backwards compatibility — they will take precedence over the UI.
